CardTable

+++ src/hotspot/cpu/x86/gc/g1/g1BarrierSetAssembler_x86.cpp
@@ -308,18 +308,18 @@ void G1BarrierSetAssembler::g1_write_barrier_post(MacroAssembler* masm,
   __ movptr(cardtable, (intptr_t)ct->card_table()->byte_map_base());
   __ addptr(card_addr, cardtable);
 
-  __ cmpb(Address(card_addr, 0), (int)G1CardTable::g1_young_card_val());
+  __ cmpb(Address(card_addr, 0), G1CardTable::g1_young_card_val());
   __ jcc(Assembler::equal, done);
 
   __ membar(Assembler::Membar_mask_bits(Assembler::StoreLoad));
-  __ cmpb(Address(card_addr, 0), (int)G1CardTable::dirty_card_val());
+  __ cmpb(Address(card_addr, 0), G1CardTable::dirty_card_val());
   __ jcc(Assembler::equal, done);
 
 
   // storing a region crossing, non-NULL oop, card is clean.
   // dirty card and log.
 
-  __ movb(Address(card_addr, 0), (int)G1CardTable::dirty_card_val());
+  __ movb(Address(card_addr, 0), G1CardTable::dirty_card_val());
 
   __ movptr(tmp2, queue_index);
   __ testptr(tmp2, tmp2);
@@ -533,17 +533,17 @@ void G1BarrierSetAssembler::generate_c1_post_barrier_runtime_stub(StubAssembler*
 
   NOT_LP64(__ get_thread(thread);)
 
-  __ cmpb(Address(card_addr, 0), (int)G1CardTable::g1_young_card_val());
+  __ cmpb(Address(card_addr, 0), G1CardTable::g1_young_card_val());
   __ jcc(Assembler::equal, done);
 
   __ membar(Assembler::Membar_mask_bits(Assembler::StoreLoad));
-  __ cmpb(Address(card_addr, 0), (int)CardTable::dirty_card_val());
+  __ cmpb(Address(card_addr, 0), CardTable::dirty_card_val());
   __ jcc(Assembler::equal, done);
 
   // storing region crossing non-NULL, card is clean.
   // dirty card and log.
 
-  __ movb(Address(card_addr, 0), (int)CardTable::dirty_card_val());
+  __ movb(Address(card_addr, 0), CardTable::dirty_card_val());
 
   const Register tmp = rdx;
   __ push(rdx);
